training/requirements.txt
torch>=2.4.0
transformers>=4.45.0
trl>=0.15.0
peft>=0.13.0
accelerate>=1.0.0
datasets>=3.0.0
bitsandbytes>=0.44.0
flash-attn>=2.6.0
pillow>=10.0.0
pyyaml>=6.0
wandb>=0.18.0
requests>=2.31.0
